refactor(daemon): extract skills routes into standalone module
Move ~470 LOC of skills API routes, types, helpers, and cache state
from daemon.ts into routes/skills.ts. Uses lazy getters for AGENTS_DIR
and SKILLS_DIR so env overrides take effect at call time (testable).
Adds 23 tests covering frontmatter parsing, install formatting,
listInstalledSkills with temp fixtures, and route integration tests
against real filesystem state.
